refactor: Separate tag and schedule into distinct domain and entity folders
