+++ noweekend-core/core-domain/src/main/kotlin/noweekend/core/domain/schedule/ScheduleReader.kt
@@ -1,4 +1,4 @@
-package noweekend.core.domain.tag
+package noweekend.core.domain.schedule
 
 import org.springframework.stereotype.Component
 import org.springframework.transaction.annotation.Transactional
