+++ noweekend-core/core-domain/src/main/kotlin/noweekend/core/domain/schedule/ScheduleRepository.kt
@@ -1,4 +1,4 @@
-package noweekend.core.domain.tag
+package noweekend.core.domain.schedule
 
 import java.time.LocalDateTime
