+++ noweekend-storage/db-core/src/main/kotlin/noweekend/storage/db/core/tag/TagHistoryCoreRepository.kt
@@ -1,4 +1,4 @@
-package noweekend.storage.db.core.schedule
+package noweekend.storage.db.core.tag
 
 import noweekend.core.domain.tag.Tag
 import noweekend.core.domain.tag.TagRepository
