+++ noweekend-storage/db-core/src/main/kotlin/noweekend/storage/db/core/tag/TagHistoryEntity.kt
@@ -1,4 +1,4 @@
-package noweekend.storage.db.core.schedule
+package noweekend.storage.db.core.tag
 
 import jakarta.persistence.Column
 import jakarta.persistence.Entity
