+++ noweekend-storage/db-core/src/main/kotlin/noweekend/storage/db/core/tag/TagRecommendCacheEntity.kt
@@ -1,4 +1,4 @@
-package noweekend.storage.db.core.cache
+package noweekend.storage.db.core.tag
 
 import jakarta.persistence.Column
 import jakarta.persistence.Entity
